Expose find_uv_bin and declare typing support
